Cleaned up this template
Added links to https://github.com/cul-it/qa_server/tree/master/config/authorities/linked_data/scenarios (where I think we want new tests to go). If we decide to lock down the repo, new branches will be created with each commit. If we decide to go with a develop branch, we'll have to update this link. May consider a README to guide the creation of YAML files.