write out cache after completing monitor tests run
The normal processing is for the monitoring tests to run once per day.  The rest of the time cached data is used to display the monitor status page.

A side effect of running the tests is that a large amount of performance data is saved in the performance cache.  This PR writes out the performance cache and starts fresh with an empty cache.