remove cache of performance year graph data
Now that the kickoff of the graph generation process is cached and calls the performance year graph job when expired, the cache of the data is no longer needed.  Everytime the job is executed, the data will need to be generated.  There is a cache of the job_id to prevent race conditions.