do not directly include linkeddata and qa gems
If these gems are dependencies instead of development_dependencies, OCLCFAST_DIRECT fails to get any data back as RDF.  Instead, it comes back as XML.  It is not clear why this happens and may not have happened outside of the .internal_test app.  But since it is not required for these to be direct dependencies, they were switched back to development_dependencies.